requirements.txt
flask>=3.0
flask-cors>=4.0
requests>=2.28
google-generativeai>=0.8.0
gunicorn>=22.0
faiss-cpu>=1.7
numpy>=1.24
sentence-transformers>=2.2
pypdf>=3.0
